# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vacancies": {
      "get": {
        "operationId": "VacanciesController_listCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/vacancies/{id}": {
      "get": {
        "operationId": "VacanciesController_getCatalog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      },
      "delete": {
        "operationId": "VacanciesController_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/vacancies/{id}/viewed": {
      "patch": {
        "operationId": "VacanciesController_viewed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/vacancies/{id}/applied": {
      "patch": {
        "operationId": "VacanciesController_applied",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/vacancies/{id}/analyze-resume": {
      "post": {
        "operationId": "VacanciesController_analyzeResume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmResumeAnalysisResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/vacancies/{id}/cover-letter": {
      "post": {
        "operationId": "VacanciesController_coverLetter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/my-vacancies": {
      "get": {
        "operationId": "MyVacanciesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyListDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "my-vacancies"
        ]
      },
      "post": {
        "operationId": "MyVacanciesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyCreateBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "my-vacancies"
        ]
      }
    },
    "/api/my-vacancies/{id}": {
      "put": {
        "operationId": "MyVacanciesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyUpdateBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredVacancyRowDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "my-vacancies"
        ]
      },
      "delete": {
        "operationId": "MyVacanciesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "my-vacancies"
        ]
      }
    },
...
    "/api/admin/vacancies/{id}": {
      "delete": {
        "operationId": "AdminVacanciesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin-vacancies"
        ]
      }
    },
...
  },
  "info": {
    "title": "job-research API",
    "description": "REST API локального ассистента поиска работы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "requestId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "formErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "StoredVacancyListDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            },
            "source": {
              "type": "string",
              "enum": [
                "manual",
                "trudvsem",
                "superjob"
              ]
            },
            "externalId": {
              "type": "string",
              "nullable": true
            },
            "ownerUserId": {
              "type": "number",
              "nullable": true
            },
            "data": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "employer": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": [
                    "name"
                  ]
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "regionCode": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "name"
                  ]
                },
                "description": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "salary": {
                  "type": "object",
                  "properties": {
                    "from": {
                      "type": "number",
                      "nullable": true
                    },
                    "to": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string"
                    },
                    "gross": {
                      "type": "boolean",
                      "nullable": true
                    }
                  },
                  "required": [
                    "from",
                    "to",
                    "currency"
                  ],
                  "nullable": true
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "publishedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "title",
                "employer",
                "location",
                "description",
                "url",
                "publishedAt"
              ]
            },
            "isPublished": {
              "type": "boolean"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "userState": {
              "type": "object",
              "properties": {
                "isViewed": {
                  "type": "boolean"
                },
                "isApplied": {
                  "type": "boolean"
                },
                "hidden": {
                  "type": "boolean"
                },
                "score": {
                  "type": "number",
                  "nullable": true
                },
                "scoreReason": {
                  "type": "string",
                  "nullable": true
                },
                "coverLetter": {
                  "type": "string",
                  "nullable": true
                },
                "processedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                }
              },
              "required": [
                "isViewed",
                "isApplied",
                "hidden",
                "score",
                "scoreReason",
                "coverLetter",
                "processedAt"
              ]
            }
          },
          "required": [
            "id",
            "source",
            "externalId",
            "ownerUserId",
            "data",
            "isPublished",
            "createdAt"
          ]
        }
      },
      "StoredVacancyRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "trudvsem",
              "superjob"
            ]
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "ownerUserId": {
            "type": "number",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "employer": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "required": [
                  "name"
                ]
              },
              "location": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "regionCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              },
              "description": {
                "type": "string"
              },
              "url": {
                "type": "string",
                "format": "uri"
              },
              "salary": {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "number",
                    "nullable": true
                  },
                  "to": {
                    "type": "number",
                    "nullable": true
                  },
                  "currency": {
                    "type": "string"
                  },
                  "gross": {
                    "type": "boolean",
                    "nullable": true
                  }
                },
                "required": [
                  "from",
                  "to",
                  "currency"
                ],
                "nullable": true
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "publishedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "title",
              "employer",
              "location",
              "description",
              "url",
              "publishedAt"
            ]
          },
          "isPublished": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userState": {
            "type": "object",
            "properties": {
              "isViewed": {
                "type": "boolean"
              },
              "isApplied": {
                "type": "boolean"
              },
              "hidden": {
                "type": "boolean"
              },
              "score": {
                "type": "number",
                "nullable": true
              },
              "scoreReason": {
                "type": "string",
                "nullable": true
              },
              "coverLetter": {
                "type": "string",
                "nullable": true
              },
              "processedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            },
            "required": [
              "isViewed",
              "isApplied",
              "hidden",
              "score",
              "scoreReason",
              "coverLetter",
              "processedAt"
            ]
          }
        },
        "required": [
          "id",
          "source",
          "externalId",
          "ownerUserId",
          "data",
          "isPublished",
          "createdAt"
        ]
      },
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "LlmResumeAnalysisResponseDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 100,
            "exclusiveMaximum": false
          },
          "reason": {
            "type": "string"
          },
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "score",
          "reason"
        ]
      },
      "VacancyCreateBodyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "employerName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "locationName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50000
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "salaryFrom": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "salaryTo": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "salaryCurrency": {
            "type": "string",
            "maxLength": 10
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "string",
              "maxLength": 100
            }
          }
        },
        "required": [
          "title",
          "employerName",
          "locationName",
          "description",
          "url"
        ]
      },
...
    }
  }
}